+++ pympc/__init__.py
@@ -1,10 +1,10 @@
 from .pympc import (
     minor_planet_check,
-    _minor_planet_check,
     update_catalogue,
-    _cone_search_xephem_entries,
     planet_hill_sphere_check,
-    _planet_hill_sphere_check,
+)
+from .utils import (
+    add_logging,
 )